+++ provider/pro/flatten.go
@@ -3,8 +3,9 @@ package pro
 import (
 	"context"
 	"fmt"
-	"github.com/RedisLabs/terraform-provider-rediscloud/provider/client"
 	"strconv"
+
+	"github.com/RedisLabs/terraform-provider-rediscloud/provider/client"
 
 	"github.com/RedisLabs/rediscloud-go-api/redis"
 	"github.com/RedisLabs/rediscloud-go-api/service/databases"
@@ -99,6 +100,11 @@ func FlattenCloudDetails(cloudDetails []*subscriptions.CloudDetail, isResource b
 			"cloud_account_id": strconv.Itoa(redis.IntValue(currentCloudDetail.CloudAccountID)),
 			"region":           regions,
 		}
+
+		if currentCloudDetail.AWSAccountID != nil {
+			cdlMapString["aws_account_id"] = redis.StringValue(currentCloudDetail.AWSAccountID)
+		}
+
 		cdl = append(cdl, cdlMapString)
 	}