+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
+# 2.8.0 (10th November 2025)
+
+## Added
+- Added support for database version for Essentials databases.
+- Added `aws_account_id` attribute to Pro and Active-Active subscription resources and data sources.
+- Added `region_id` to the attribute reference documentation for `rediscloud_active_active_subscription_regions` data source.
+- Added `region_id` attribute to `rediscloud_regions` data source.
+- Added `db_id` to the attribute reference documentation for `rediscloud_database` data source.
+
+## Fixed
+- Spurious diffs for `customer_managed_key_deletion_grace_period` are now suppressed when upgrading the provider.
 
 # 2.7.4 (7th November 2025)